Split Sample interface and class
