+++ dependencykit/src/test/java/com/github/kazy1991/dependencykit/SmartPhone.java
@@ -1,0 +1,13 @@
+package com.github.kazy1991.dependencykit;
+
+
+public interface SmartPhone {
+
+    String developer();
+
+    String os();
+
+    int storageSize();
+
+    long weight();
+}
